KoinModule.kt moved to di package
